pr: self-review
