+++ plugins/dependencies/src/main/kotlin/de/fayard/refreshVersions/VersionsCatalogTask.kt
@@ -1,4 +1,4 @@
-package de.fayard.refreshVersions.internal
+package de.fayard.refreshVersions
 
 import de.fayard.refreshVersions.core.addMissingEntriesInVersionsProperties
 import de.fayard.refreshVersions.core.internal.Case
@@ -9,6 +9,8 @@ import de.fayard.refreshVersions.core.internal.OutputFile
 import de.fayard.refreshVersions.core.internal.checkModeAndNames
 import de.fayard.refreshVersions.core.internal.computeUseFqdnFor
 import de.fayard.refreshVersions.core.internal.findDependencies
+import de.fayard.refreshVersions.internal.Toml.versionsCatalog
+import de.fayard.refreshVersions.internal.getArtifactNameToConstantMapping
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.tasks.TaskAction
@@ -66,55 +68,6 @@ open class VersionsCatalogTask : DefaultTask() {
     companion object {
         val currentGradleVersion = GradleVersion.current()
         val versionWithVersionsCatalog = GradleVersion.version("7.4")
-
-        internal fun parseTomlInSection(toml: String): Map<String, String> {
-            val result = mutableMapOf<String, StringBuilder>()
-            result["root"] = StringBuilder()
-            var current: StringBuilder = result["root"]!!
-            val lines = toml.lines()
-            for ((index, line) in lines.withIndex()) {
-                val trimmed = line.trim()
-                val isSectionHeader = trimmed.startsWith("[") && trimmed.endsWith("]")
-                if (isSectionHeader) {
-                    val sectionName = trimmed.removePrefix("[").removeSuffix("]")
-                    result[sectionName] = StringBuilder()
-                    current = result[sectionName]!!
-                } else {
-                    current.append(line)
-                    if (index != lines.lastIndex) current.append("\n")
-                }
-            }
-            return result.mapValues { it.value.toString() }
-        }
-
-        internal fun tomlSectionsToString(sections: Map<String, String>): String = buildString {
-            for ((header, content) in sections) {
-                if (header != "root") append("[$header]\n")
-                append(content)
-            }
-        }
-
-        internal fun versionsCatalog(deps: Deps, currentText: String): String {
-            val sections = parseTomlInSection(currentText).toMutableMap()
-            if (sections["root"].isNullOrBlank()) {
-                sections["root"] = "## Generated by $ ./gradlew refreshVersionsCatalog\n\n"
-            }
-            sections["libraries"] = versionsCatalogLibraries(deps)
-            return tomlSectionsToString(sections)
-        }
-
-        internal fun versionsCatalogLibraries(deps: Deps) = buildString {
-            append('\n')
-            deps.libraries.forEach {
-                append(deps.names[it])
-                append(" = \"")
-                append(it.groupModuleUnderscore())
-                append('"')
-                append("\n")
-                append("\n")
-            }
-            append("\n")
-        }
     }
 }
