+++ plugins/dependencies/src/main/kotlin/de/fayard/refreshVersions/internal/Toml.kt
@@ -1,0 +1,55 @@
+package de.fayard.refreshVersions.internal
+
+import de.fayard.refreshVersions.core.internal.Deps
+
+internal object Toml {
+
+    fun parseTomlInSection(toml: String): Map<String, String> {
+        val result = mutableMapOf<String, StringBuilder>()
+        result["root"] = StringBuilder()
+        var current: StringBuilder = result["root"]!!
+        val lines = toml.lines()
+        for ((index, line) in lines.withIndex()) {
+            val trimmed = line.trim()
+            val isSectionHeader = trimmed.startsWith("[") && trimmed.endsWith("]")
+            if (isSectionHeader) {
+                val sectionName = trimmed.removePrefix("[").removeSuffix("]")
+                result[sectionName] = StringBuilder()
+                current = result[sectionName]!!
+            } else {
+                current.append(line)
+                if (index != lines.lastIndex) current.append("\n")
+            }
+        }
+        return result.mapValues { it.value.toString() }
+    }
+
+    fun tomlSectionsToString(sections: Map<String, String>): String = buildString {
+        for ((header, content) in sections) {
+            if (header != "root") append("[$header]\n")
+            append(content)
+        }
+    }
+
+    fun versionsCatalog(deps: Deps, currentText: String): String {
+        val sections = parseTomlInSection(currentText).toMutableMap()
+        if (sections["root"].isNullOrBlank()) {
+            sections["root"] = "## Generated by $ ./gradlew refreshVersionsCatalog\n\n"
+        }
+        sections["libraries"] = versionsCatalogLibraries(deps)
+        return tomlSectionsToString(sections)
+    }
+
+    fun versionsCatalogLibraries(deps: Deps) = buildString {
+        append('\n')
+        deps.libraries.forEach {
+            append(deps.names[it])
+            append(" = \"")
+            append(it.groupModuleUnderscore())
+            append('"')
+            append("\n")
+            append("\n")
+        }
+        append("\n")
+    }
+}
