+++ sample-multi-modules/buildSrc/src/main/kotlin/Libs.kt
@@ -14,8 +14,6 @@ import kotlin.String
 object Libs {
     const val clikt: String = "com.github.ajalt.clikt:clikt:_"
 
-    const val okio: String = "com.github.javadev:okio:_"
-
     const val mockito_kotlin: String = "com.nhaarman.mockitokotlin2:mockito-kotlin:_"
 
     const val okhttp: String = "com.squareup.okhttp3:okhttp:_"
