+++ sample-multi-modules/module1/build.gradle.kts
@@ -15,5 +15,4 @@ useDependency()
 dependencies {
     implementation(Square.okHttp3.okHttp)
     testImplementation("junit", "junit", "4.12")
-    implementation("com.github.javadev:okio:1.6.0")
 }
