+++ sample-multi-modules/settings.gradle.kts
@@ -32,5 +32,3 @@ refreshVersions {
 
 include("module1")
 include("module2")
-
-enableFeaturePreview("VERSION_CATALOGS")
